adds, updates, fixes

+++ src/main/java/dik/library/dao/author/AuthorRowMapper.java
@@ -11,7 +11,7 @@ import java.sql.SQLException;
 public class AuthorRowMapper implements RowMapper<Author> {
     @Override
     public Author mapRow(ResultSet resultSet, int i) throws SQLException {
-        return new Author(resultSet.getInt("id"),
+        return new Author(resultSet.getLong("id"),
                 resultSet.getString("firstName"),
                 resultSet.getString("secondName"));
     }
